backend/requirements.txt
fastapi>=0.104.1,<0.105.0
uvicorn>=0.24.0,<0.25.0
motor>=3.3.1,<3.4.0
pydantic>=2.7.4,<3.0.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.6,<0.1.0
python-dotenv>=1.0.0,<1.1.0
pytest>=7.4.3,<7.5.0
httpx>=0.25.1,<0.26.0
python-dateutil>=2.8.2,<2.9.0
pymongo>=4.6.0,<4.7.0
bcrypt>=4.0.1,<4.1.0
email-validator>=2.1.0,<2.2.0 